+++ demo/DemoApp.vue
@@ -3,6 +3,7 @@ import { RrgPlaybackControls } from '../src'
 import DemoChartHost from './DemoChartHost.vue'
 import DemoControls from './DemoControls.vue'
 import DemoCopyOverrides from './DemoCopyOverrides.vue'
+import DemoPerfPanel from './DemoPerfPanel.vue'
 import { useDemoAppState } from './useDemoAppState'
 import { partialCopyFromFields } from './demoCopyFields'
 import './DemoApp.css'
@@ -52,6 +53,7 @@ const {
     />
 
     <DemoCopyOverrides v-model="controls" />
+    <DemoPerfPanel />
 
     <p v-if="copyStatus" class="copy-status" data-testid="demo-copy-status">{{ copyStatus }}</p>